Replace Sentry.Util with Sentry.UUID
